#7 Properly html-escape issue title
Previously the whole issue was passed to the escape function. The escape function tried to convert the issue to string, which was not implicitly possible and caused the error `TypeError (no implicit conversion of Issue into String):`. By calling to_s explicitly, we prevent this error. This leads into using the issue title as message.